Widget class for a Spectrum and a Baseline.
The GuiABS was changed in a WidgetBase class. This class has many
widgets and properties one most likely needs for any future widget
that handles data based on AllYouCanEat data.

AllYouCanPlot changed to SpecAndSummed.

Added SpecAndBase class to have a gui handler for subtracting baseline
frome data. Ideally the same gui can be used for IR and the PumpProbe
Spectra.