Tabed Dasboard added
The tabed dasboard has two pages by now. The first shows signal and
sumed signal. And allows for optional baseline substraction.

The second Page shows signal and Baseline.

The Idear is, that the first Page is used for the pump-probe scan
and the second page for the IR-Profile.

Maybe on a following The Normalized signal will then be shown on either
the first or a 3rd page. I have not decided yet.