Possibility to select multiple spectra for baseline subtraction
The spectrum slider of the baseline in now a RangeSlider, that allows
to select multiple spectra from the baseline. When ever there is more
then one baseline spectrum selected, the baseline correction will
apply between the complete data set of the raw spectra and the baseline.
In the case of having selected data with unmatchable shaped, a warning
will be raised, and the subtract baseline button will be reset to
false.